Add files via upload
The program uses the cosine similarity metric to recommend movies based on their genre and overview. It creates a vector representation of each movie's tags using the CountVectorizer and then calculates the cosine similarity between each pair of movie vectors. The recommendator function takes a movie title as input and returns the titles of the 5 most similar movies.

The results show that the program is able to recommend movies that are similar in genre and theme to the input movie. For example, when "Iron Man" is inputted, the program recommends other superhero movies such as "The Avengers" and "Captain America: Civil War". When "The Godfather" is inputted, the program recommends other crime dramas such as "The Godfather Part II" and "Goodfellas".

Overall, the program is a successful implementation of the cosine similarity algorithm for recommending movies. It is able to provide users with a list of movies that are similar to their favorite movies, which can help them discover new movies to watch.